refactor: move controller directory to infrastructure dir
following clean architecture, controller is part of the infrastructure
layer
