+++ src/main/java/com/training/todo/infrastructure/controller/servlet/TaskServlet.java
@@ -1,8 +1,8 @@
-package com.training.todo.controller.servlet;
+package com.training.todo.infrastructure.controller.servlet;
 
 import com.training.todo.application.exceptions.InvalidTaskException;
-import com.training.todo.controller.model.TaskDto;
-import com.training.todo.controller.service.TaskService;
+import com.training.todo.infrastructure.controller.model.TaskDto;
+import com.training.todo.infrastructure.controller.service.TaskService;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletConfig;
 import jakarta.servlet.ServletContext;
