+++ src/main/webapp/todoForm/form.jsp
@@ -1,4 +1,4 @@
-<%@ page import="com.training.todo.controller.model.TaskDto" %>
+<%@ page import="com.training.todo.infrastructure.controller.model.TaskDto" %>
 <%@ page contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
 <!DOCTYPE html>
 <html>
